Revert "[#131] api 파일 정리"
This reverts commit 3bca2875fb8673810d8cfcd105195f8e90157e07.
