+++ src/app/dashboard/[id]/components/TaskCardContent.tsx
@@ -1,5 +1,3 @@
-import Image from "next/image";
-
 import Close from '/public/icons/close.svg'
 import Kebab from '/public/icons/kebab-menu.svg'
 import ColumnChip from '@/app/dashboard/[id]/components/ColumnChip'
@@ -84,7 +82,7 @@ export default function TaskCardContent({
             {description}
           </p>
           <div className='relative mb-6 rounded-md'>
-            <Image className='rounded-md' src={imageUrl} alt='이미지' />
+            <img className='rounded-md' src={imageUrl} alt='이미지' />
           </div>
           <Form
             formId='commentForm'
