+++ src/app/login/utils/utils.ts
@@ -1,9 +1,9 @@
 import type { LoginFormValue } from '@/app/login/components/LoginForm'
-import taskifyApi from '@/lib/axiosInstance'
+import api from '@/app/utils/axiosInstance'
 
 export const login = async (data: LoginFormValue) => {
   try {
-    const response = await taskifyApi.post('auth/login', data)
+    const response = await api.post('auth/login', data)
     const { accessToken, user } = response.data
     return { accessToken, user }
   } catch (error) {
@@ -13,7 +13,7 @@ export const login = async (data: LoginFormValue) => {
 
 export const fetchDashboards = async () => {
   try {
-    const response = await taskifyApi.get(
+    const response = await api.get(
       'dashboards?navigationMethod=infiniteScroll&page=1&size=10'
     )
     const { dashboards } = response.data
